server/requirements.txt
Flask==3.0.0
flask-mysql==1.5.2
flask_sqlalchemy==3.1.1
Flask-RESTful==0.3.10
Flask-Swagger-UI==4.11.1
cryptography==42.0.5
Flask-Migrate==4.0.5
alembic==1.13.1
mysql-connector-python==8.3.0
flask_cors==4.0.0
bcrypt==4.1.2
Flask-JWT-Extended==4.6.0
sendgrid==6.11.0
pytest==8.0.2
Flask-JWT-Extended==4.6.0
pandas==2.2.1
numpy==1.24.3
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.1